AnalyserModel::areEquivalentVariables(): make sure that the key for the cache is unique.

+++ src/analysermodel_p.h
@@ -59,7 +59,7 @@ struct AnalyserModel::AnalyserModelImpl
     bool mNeedAcschFunction = false;
     bool mNeedAcothFunction = false;
 
-    std::map<intptr_t, bool> mCachedEquivalentVariables;
+    std::map<uint_fast64_t, bool> mCachedEquivalentVariables;
 };
 
 } // namespace libcellml
